updated ui styling for table data

+++ batch-processor/backend/app/models/transaction.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import datetime, timezone
+from decimal import Decimal
 from sqlalchemy import String, Boolean, Text, DateTime, ForeignKey, Numeric
 from sqlalchemy.orm import Mapped, mapped_column
 from sqlalchemy.dialects.postgresql import UUID, JSONB
@@ -12,7 +13,7 @@ class Transaction(Base):
     job_id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), ForeignKey("jobs.id", ondelete="CASCADE"), nullable=False)
     transaction_id: Mapped[str] = mapped_column(Text, nullable=False)
     user_id: Mapped[str] = mapped_column(Text, nullable=False)
-    amount: Mapped[float | None] = mapped_column(Numeric(15, 2), nullable=True)
+    amount: Mapped[Decimal | None] = mapped_column(Numeric(15, 2), nullable=True)
     timestamp: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
     is_valid: Mapped[bool] = mapped_column(Boolean, default=True)
     is_suspicious: Mapped[bool] = mapped_column(Boolean, default=False)
